Move parser into separate directory
