+++ src/Snippet.php
@@ -39,7 +39,7 @@ class Snippet extends Element\Group
         $tokens = (new Lexer\Document())($stream);
 
         // parse and render the tokens
-        $group = (new Renderer())->render((new Parser())->parse($tokens));
+        $group = (new Renderer())->render((new Parser\Document())->parse($tokens));
 
         // set the snippet's properties from the group
         $this->parent     = null;
